requirements.txt
discord.py
pillow
schedule
tweepy